Fixed TrcTrackSet::getTrackSetNumber
Discovered that TrcTrackSet::getTrackSetNumber was always returning an error. Changed it to use Uint32 instead of Uint16, as the TrackSetNumber tag has a VR of UL. Local tests showed this fix to now properly return the value.

+++ dcmtract/include/dcmtk/dcmtract/trctrackset.h
@@ -123,7 +123,7 @@ public:
     *  @param  pos Index of the value to get (0..vm-1)
     *  @return EC_Normal if successful, an error code otherwise
     */
-  virtual OFCondition getTrackSetNumber(Uint16 &value,
+  virtual OFCondition getTrackSetNumber(Uint32 &value,
                                         const unsigned long pos = 0) const;
 
   /** Get Track Set Label
